feature: user registration and upload to database
User fills out form with name, email, and password. This gets
populated into a database table called User that enforces
referential integrity to make sure no name or email is duplicated.
Currently we do not validate if the email is a valid email or not.

Along the way, we made many errors and discoveries. For example,
committing a database transaction on invalid inputs can lead to
users existing in a database even though the screen says they
failed to register, so it's very important to put try/except
statements before committing to database. This prevents situations
international students applying for work VISA's not being able
to get into their account because their name and email already
exists from the first registration page which said registration
failed, but the commit to the database still occurred.

We haven't implemented the user login yet, but it should be
similar to registration. We also have not encrpyted the password.