feature: play a song by submitting request form with song_id
Play song feature only supports .mp3 files right now. There are
known issues with the `init_songs.py` script that can be read in
detail within the README.md. Also we cannot figure out how to
make a song playable with a click button; instead we have to submit
a form with the song ID as input.