feature: add user to db via GET request call
To create a REST API that supports GET requests, we must add an
if/else statement to the route function that maps to the url of
the GET request. The if else statement must explicitly say how
it will handle each type of HTTP request. For example if it only
handles GET requests, then when `request.method`=="GET" we can
try to insert a new user to db. Otherwise if it is not a GET
request, then we immediately send a response saying the request
is not valid.